+++ src/components/AreaChart/AreaChart.tsx
@@ -1,4 +1,4 @@
-// Tremor Raw AreaChart [v0.3.1]
+// Tremor AreaChart [v0.3.1]
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 "use client"