+++ src/components/Checkbox/Checkbox.tsx
@@ -1,4 +1,4 @@
-// Tremor Raw Checkbox [v0.0.3]
+// Tremor Checkbox [v0.0.3]
 
 import React from "react"
 import * as CheckboxPrimitives from "@radix-ui/react-checkbox"