+++ src/components/LineChart/LineChart.tsx
@@ -1,4 +1,4 @@
-// Tremor Raw LineChart [v0.3.2]
+// Tremor LineChart [v0.3.2]
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 "use client"